move `CollectionTree` into main source as a new module `pytest_helpers`
